requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiogram-dialog==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
environs==11.0.0
frozenlist==1.4.1
greenlet==3.1.1
h3==3.7.7
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
nats-py==2.9.0
numpy==2.1.2
packaging==24.1
pluggy==1.5.0
psycopg==3.1.19
psycopg-binary==3.1.19
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
setuptools==75.1.0
SQLAlchemy==2.0.32
timezonefinder==6.5.3
typing_extensions==4.12.2
yarl==1.9.4
